requirements_file.txt
# Core LangChain and LangGraph
langchain>=0.1.0
langgraph>=0.0.40
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# UI Framework
streamlit>=1.28.0

# Vector Database and Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.2

# Document Processing
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
pandas>=1.5.0

# Web Search and Tools
duckduckgo-search>=3.9.0
requests>=2.31.0

# Environment Management
python-dotenv>=1.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Optional: Advanced NLP features
# transformers>=4.30.0
# torch>=2.0.0

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Additional tools (uncomment if needed)
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
wikipedia>=1.4.0
wolframalpha>=5.0.0
